feat:完善点位搜索

+++ lkd-server/src/main/java/com/lkd/mapper/RegionMapper.java
@@ -5,7 +5,6 @@ import com.lkd.vo.RegionVO;
 import org.apache.ibatis.annotations.Insert;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
-import org.apache.ibatis.annotations.Update;
 
 import java.util.List;
 
@@ -23,5 +22,5 @@ public interface RegionMapper {
 
     //区域查询
     @Select("select * from tb_region where id = #{regionId}")
-    RegionVO findById(Integer regionId);
+    RegionVO findById(Long regionId);
 }
